pyproject.toml
[project]
name = "laptop-rag-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "lxml>=6.0.2",
    "numpy>=2.4.4",
    "playwright>=1.58.0",
    "requests>=2.33.1",
    "sentence-transformers>=5.3.0",
    "torch>=2.11.0",
    "tqdm>=4.67.3",
]
